+++ @es-labs/node/config.default.js
@@ -71,6 +71,22 @@ global.CONFIG.KNEXFILE = null // knexfile
 global.CONFIG.GCP_SERVICE_KEY = null // GCP SERVICE KEY { }
 global.CONFIG.GCP_DEFAULT_BUCKET = 'mybot-live.appspot.com'
 
+// helmetjs options
+global.CONFIG.HELMET_OPTIONS = {
+  hideServer: true,
+  csp: true,
+  nosniff: true,
+  xssfilter: true,
+  // csp: {
+  //   directives: {
+  //     defaultSrc: ["'self'"],
+  //     scriptSrc: ["'self'",'code.jquery.com','maxcdn.bootstrapcdn.com'],
+  //     styleSrc: ["'self'",'maxcdn.bootstrapcdn.com'],
+  //     fontSrc: ["'self'",'maxcdn.bootstrapcdn.com']
+  //   }
+  // }
+}
+
 // CORS - SAME ORIGIN - PROXIED
 //   CORS_OPTIONS: null
 //   PROXY_WWW_ORIGIN: 'example.com:8080'